Handle empty response deserialisation correctly
This is complicated by the empty response (or string) not being valid
JSON; so serde_json somewhat correctly fails. Somehow we need to
distinguish empty responses (which should deserialise to a unit type)
from other classes of invalid JSON, for which we construct the
`MalformedJSON` variant holding the invalid body as a String.

The solution in this commit is to wrap the deserialisation with a
function generic over deserialized types; which instead deserialises
`null` if the response is empty, resulting in the correct type `()`.
This function, `deser_into`, returns an `Err` with the Stringified body
if both attempts fail: the empty response could still be unexpected, and
itself an instance of `MalformedJSON`.