Move resource ID if DELETE request
This commit changes request function signatures to take all path
parameters by reference - except the last one, if it is a DELETE
request. That is, for a request to:

    /collection/{ID}

the signatures are:

    fn get(&T) -> ResponseResult
    // ...
    fn delete(T) -> ResponseResult

while for:

    /collection/{IDa}/subcollection/{IDb}

they are:

    fn get(&T, &U) -> ResponseResult
    // ...
    fn delete(&T, U) -> ResponseResult

This is to prevent further requests to a DELETEd resource; the compiler
will now complain that `IDb` was moved into `delete`, so the later
attempt to borrow it is not allowed.