Use placeholder for valid header when no auth
This is silly, (!FIXME) but with the current architecture it'd be
hard/messy to not have an auth argument at all when there's no security
requirements object at the root level of the schema.

If the whole `infer` module were refactored to be serde
visitors and deserialisers, then this could be more neatly achieved.